core: Create BaseParser
